+++ src/main/java/store/readData/ProductFileReader.java
@@ -1,4 +1,4 @@
-package store.dataBase;
+package store.readData;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -7,12 +7,12 @@ import java.util.List;
 import store.constant.Error;
 import store.constant.File;
 
-public class PromotionFileReader {
+public class ProductFileReader {
 
-    public List<String> readPromotion() {
+    public List<String> readProduct() {
         while (true) {
             try {
-                List<String> input = Files.readAllLines(Paths.get(File.PROMOTION_FILE_PATH));
+                List<String> input = Files.readAllLines(Paths.get(File.PRODUCT_FILE_PATH));
                 input.remove(File.COLUMN_HEADERS);
                 return input;
             } catch (IOException e) {
