+++ src/main/java/store/service/StoreService.java
@@ -5,8 +5,8 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import store.dataBase.ProductFileReader;
-import store.dataBase.PromotionFileReader;
+import store.readData.ProductFileReader;
+import store.readData.PromotionFileReader;
 import store.constant.Error;
 import store.model.Product;
 import store.model.ShoppingItem;
